script to allow to remove the output data produced by a production. this is a subset of the functionality provided by dirac-production-clean-production.
using the directories found in the production management system it obtain all the files from the lfc and removes them from the storage and catalogs.
the same directories are listed on the storages to pick up any remnants which are then removed.
finally, all the files found in the BK are removed.


git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/LHCbDIRAC/trunk@17285 4525493e-7705-40b1-a816-d608a930855b